Throw on missing keys
Throwing an exception enables clients to detect failures to render
template completely due to missing keys. This is useful in cases where
clients expect templates and the data structure to always match and
treat missing keys as failures.

Define a custom exception to enable clients to identify the specific
type of the error.

Signed-off-by: Hannu Lounento <hannu.lounento@vaisala.com>